requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
grpcio==1.48.2
pandas
openpyxl
python-dotenv
azure-storage-blob
azure-identity
azure-keyvault-secrets

